+++ shenyu-plugin-auth/src/main/resources/application.properties
