+++ shenyu-spring-boot-starter-plugin-auth/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.apache.shenyu.springboot.starter.plugin.auth.AuthPluginConfiguration
