+++ shenyu-spring-boot-starter-plugin-auth/src/main/resources/META-INF/spring.provides
@@ -1,0 +1,1 @@
+provides: shenyu-spring-boot-starter-plugin-auth
